moved header component out of shared, since it is only used by job list
