+++ ui/src/app/app.component.css
@@ -14,6 +14,11 @@
   font-weight: 400;
 }
 
+::ng-deep .mat-icon-button,
+::ng-deep .mat-raised-button {
+  background-color: inherit;
+}
+
 ::ng-deep a {
   font-weight: 500;
 }
