+++ ui/src/app/job-list/header/chips/filter-chip.component.css
