+++ ui/src/app/job-list/header/chips/status-selection.component.ts
@@ -1,4 +1,4 @@
-import {JobStatus} from "../../model/JobStatus";
+import {JobStatus} from "../../../shared/model/JobStatus";
 import {Component, EventEmitter, Input, OnInit, Output} from "@angular/core";
 
 @Component({
